[User/libc] Syscalls get `sys_...` wrappers
My hope is that this will introduce at least *some* semblance of type
safety when it comes to using syscalls. Plus, this file already needs
updated when making a new syscall. It also consolidates the definition
of some structs like `sockaddr` and `tm`, since those are needed for
the syscall interface. We may want to move those into `sys/types.h`,
but `sys/types.h` is a POSIX header and we don't want to mess with
that too much.