[Kernel] Expand upon events (add `EventFilter`)
This will allow an event queue to listen to specific process file
descriptors and whether or not they are ready to read to or write
from.

Still thinking about the best way to do things, but this doesn't seem
too bad; at least, it should hopefully work and not be *terribly*
slow. But it will be slower than it could be; there are certainly
multiple things that could be optimised.