[Kernel] More progress on events...
It looks like we are actually able to have events happen and
everything work! HOWEVER! If you run LensorOS and the *first* thing
you do is `bg servertest RET clienttest RET` then it causes a page
fault where the kernel dereferences NULL through a virtual address
of (0xffffffff0000000)... so no idea how that happens but if you run
another process *and then* try to do the server/client test it works.
So yeah, something to work on there...