[Bootloader] Fixup makefile, I think
Because `--fatal-warnings` is passed to ld from *somewhere*, and newer
versions of ld warn on executable stack, that caused it to error every
time. On top of that, it wasn't looking in the right place for the
object files and libraries it was linking with, so I fixed those paths
up (I think).

If you are reading this from a bisect because you can no longer build
the bootloader, I am dearly sorry.