[BEGIN] Begin to implement changes to move input, pipes, sockets to FilesystemDriver
Previously, they were just StorageDeviceDrivers. But, I think the new
squeeze is, we will have the VFS only interact with FilesystemDrivers,
rather than the current "tunneling" from the VFS to the storage device.
This is necessary but just hasn't been done yet due to the sheer size
of the change. Everything will be broken for a long, long time.