[FAT] Don't return volume IDs/labels as files
You know how when you plug a USB drive in it has a name that pops up?
That name is stored in the root directory in a volume label directory
entry in the FAT. Our directory_data implementation was including this
volume label as a regular file, when it really isn't one; this changes
it to skip past any and all volume labels it sees, so as to only
return regular files and subdirectories.