[Scheduler] Do not free page tables
Basically, I can't figure out how to do this properly. Doing it in the
kernel thread like we were before causes iterator invalidation across
yields.

I think what we may want to do is, in the main loop in the kernel
thread, to check for a threshold amount of page maps that need freed,
and, if it has been passed, DISABLE INTERRUPTS (no yield), free them
all, then reenable interrupts to allow for a yield away.