[libc] Changes to stdio implementation
Basically, I'm not sure if any of this actually:
  A.) changes anything, and
  B.) doesn't break anything.

So, there may be a terrible bug caused by these changes, or one
fixed. These changes were part of figuring out that files have sizes
and the read syscall/VFS never dealt with that. Either way, this is
what's on my machine and it's currently working well enough for me to
not feel the need to change anything back. So, if you're reading this
in the future because libc FILEs are completely borked, uhh, pretend I
had a good reason for these changes :P.