[UTF] Use `string_view` parameter instead of `string`
Since the input to `utf16_to_utf8` is readonly, we don't need it to
own the memory it points to like `string  does. Instead, we can take a
view of some memory owned elsewhere using a `string_view`. This will
remove the need for a memory allocation in a lot of cases.