[KSTAGE1] Better represent partition names upon discovery
Very minor, just prints partition names in a human-readable form
without a bunch of null bytes in there and without a bunch of spaces
on the end.

Also, what in the world was the `0x38` there for before?? Kind of
insane, can't even picture why I'd hardcode the size of something like
that.